add logic to check log out at the root level
