+++ frontend/src/components/Dashbard.tsx
@@ -25,18 +25,18 @@ const tab = {
 export const Dashboard = (): JSX.Element => {
 
 	// use history hook to redirect/navigate between routes
-	const history = useNavigate();
+	// const history = useNavigate();
 
 	// Redux connect equivalent
 	const dispatch = useDispatch();
 	const deleteDataset = (datasetId: string) => dispatch(datasetDeleted(datasetId));
 	const listDatasets = (skip: number | undefined, limit: number | undefined, mine: boolean | undefined) => dispatch(fetchDatasets(skip, limit, mine));
 	const dismissError = () => dispatch(resetFailedReason());
-	const dismissLogout = () => dispatch(resetLogout());
+	// const dismissLogout = () => dispatch(resetLogout());
 	const datasets = useSelector((state: RootState) => state.dataset.datasets);
 	const reason = useSelector((state: RootState) => state.error.reason);
 	const stack = useSelector((state: RootState) => state.error.stack);
-	const loggedOut = useSelector((state: RootState) => state.error.loggedOut);
+	// const loggedOut = useSelector((state: RootState) => state.error.loggedOut);
 
 	const [datasetThumbnailList, setDatasetThumbnailList] = useState<any>([]);
 	// TODO add option to determine limit number; default show 5 datasets each time
@@ -80,14 +80,14 @@ export const Dashboard = (): JSX.Element => {
 		window.open(`${config.GHIssueBaseURL}+${reason}&body=${encodeURIComponent(stack)}`);
 	}
 
-	// log user out if token expired/unauthorized
-	useEffect(() => {
-		if (loggedOut) {
-			// reset loggedOut flag so it doesn't stuck in "true" state, then redirect to login page
-			dismissLogout();
-			history("/auth/login");
-		}
-	}, [loggedOut]);
+	// // log user out if token expired/unauthorized
+	// useEffect(() => {
+	// 	if (loggedOut) {
+	// 		// reset loggedOut flag so it doesn't stuck in "true" state, then redirect to login page
+	// 		dismissLogout();
+	// 		history("/auth/login");
+	// 	}
+	// }, [loggedOut]);
 
 	// fetch thumbnails from each individual dataset/id calls
 	useEffect(() => {
