+++ frontend/src/index.tsx
@@ -4,7 +4,7 @@
 import React from "react";
 import { render } from "react-dom";
 import { Provider } from "react-redux";
-import AppRoutes from "./routes";
+import { AppRoutes } from "./routes";
 import configureStore from "./store/configureStore";
 import {ThemeProvider} from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
@@ -16,7 +16,7 @@ render(
 	<ThemeProvider theme={theme}>
 		<Provider store={store}>
 			<CssBaseline />
-			{AppRoutes}
+			<AppRoutes />
 		</Provider>
 	</ThemeProvider>
 	, document.getElementById("app")
